refactor(canopy): extract hydrateCanopyDescription to shared module
